+++ src/main/kotlin/racingcar/strategy/MovingStrategy.kt
@@ -1,0 +1,5 @@
+package racingcar.strategy
+
+interface MovingStrategy {
+    fun move(): Boolean
+}
