+++ src/main/kotlin/racingcar/util/random/RandomNumber.kt
@@ -1,0 +1,7 @@
+package racingcar.util.random
+
+class RandomNumber(
+    private val range: IntRange
+) : RandomUtil {
+    override fun generate() = range.random()
+}
