+++ src/main/kotlin/racingcar/util/random/RandomUtil.kt
@@ -1,0 +1,5 @@
+package racingcar.util.random
+
+interface RandomUtil {
+    fun generate(): Int
+}
