--- src/main/kotlin/step3_racingcar/strategy/MovingStrategy.kt
@@ -1,5 +1,0 @@
-package step3_racingcar.strategy
-
-interface MovingStrategy {
-    fun move(): Boolean
-}
