--- src/main/kotlin/step3_racingcar/util/random/RandomNumber.kt
@@ -1,7 +1,0 @@
-package step3_racingcar.util.random
-
-class RandomNumber(
-    private val range: IntRange
-) : RandomUtil {
-    override fun generate() = range.random()
-}
