--- src/main/kotlin/step3_racingcar/util/random/RandomUtil.kt
@@ -1,5 +1,0 @@
-package step3_racingcar.util.random
-
-interface RandomUtil {
-    fun generate(): Int
-}
