--- src/main/kotlin/step3_racingcar/view/ResultView.kt
@@ -1,16 +1,0 @@
-package step3_racingcar.view
-
-import step3_racingcar.Car
-
-class ResultView {
-    fun printCarDistance(cars: List<Car>) {
-        cars.forEach {
-            println(SPACE.repeat(it.distance))
-        }
-        println()
-    }
-
-    companion object {
-        private const val SPACE = "-"
-    }
-}
